WIP: refactor to abstract contract market
