+++ MekHQ/src/mekhq/campaign/market/AtbMonthlyContractMarket.java
@@ -38,7 +38,6 @@ import megamek.common.annotations.Nullable;
 import megamek.common.enums.SkillLevel;
 import megamek.logging.MMLogger;
 import mekhq.MekHQ;
-import mekhq.campaign.market.enums.ContractMarketMethod;
 import mekhq.utilities.MHQXMLUtility;
 import mekhq.campaign.Campaign;
 import mekhq.campaign.JumpPath;
@@ -64,38 +63,7 @@ import mekhq.campaign.universe.Systems;
  *
  * @author Neoancient
  */
-public class AtbMonthlyContractMarket implements IContractMarket {
-    private static final MMLogger logger = MMLogger.create(ContractMarket.class);
-
-    protected List<Contract> contracts;
-    protected int lastId = 0;
-    protected Map<Integer, Contract> contractIds;
-
-    private ContractMarketMethod method;
-
-    // TODO: Implement a method that rolls each day to see whether a new contract appears or an offer disappears
-    private Map<Integer, AtbMonthlyContractMarket.ClauseMods> clauseMods;
-
-    /**
-     * An arbitrary maximum number of attempts to generate a contract.
-     */
-    private final static int MAXIMUM_GENERATION_RETRIES = 3;
-
-    /* It is possible to call addFollowup more than once for the
-     * same contract by canceling the dialog and running it again;
-     * this is the easiest place to track it to prevent
-     * multiple followup contracts.
-     * key: followup id
-     * value: main contract id
-     */
-    private HashMap<Integer, Integer> followupContracts;
-
-    /**
-     * An arbitrary maximum number of attempts to find a random employer faction that
-     * is not a Mercenary.
-     */
-    private final static int MAXIMUM_ATTEMPTS_TO_FIND_NON_MERC_EMPLOYER = 20;
-
+public class AtbMonthlyContractMarket extends AbstractContractMarket {
     public AtbMonthlyContractMarket() {
         contracts = new ArrayList<>();
         contractIds = new HashMap<>();
@@ -104,31 +72,12 @@ public class AtbMonthlyContractMarket implements IContractMarket {
     }
 
     @Override
-    public void removeContract(Contract c) {
-        contracts.remove(c);
-        contractIds.remove(c.getId());
-        clauseMods.remove(c.getId());
-        followupContracts.remove(c.getId());
-    }
-
     public AtBContract addAtBContract(Campaign campaign) {
         AtBContract c = generateAtBContract(campaign, campaign.getAtBUnitRatingMod());
         if (c != null) {
             contracts.add(c);
         }
         return c;
-    }
-
-    @Override
-    public int getRerollsUsed(Contract c, int clause) {
-        if (null != clauseMods.get(c.getId())) {
-            return clauseMods.get(c.getId()).rerollsUsed[clause];
-        }
-        return 0;
-    }
-
-    public List<Contract> getContracts() {
-        return contracts;
     }
 
     @Override
@@ -152,11 +101,6 @@ public class AtbMonthlyContractMarket implements IContractMarket {
             clauseMods.get(c.getId()).rerollsUsed[clause]++;
             c.calculateContract(campaign);
         }
-    }
-
-    @Override
-    public void generateContractOffers(Campaign campaign) {
-        generateContractOffers(campaign, false);
     }
 
     @Override
@@ -220,11 +164,8 @@ public class AtbMonthlyContractMarket implements IContractMarket {
                 if (campaign.getAtBConfig().isHiringHall(campaign.getCurrentSystem().getId(),
                         campaign.getLocalDate())) {
                     numContracts++;
-                    /*
-                     * Though the rules do not state these modifiers are mutually exclusive, the
-                     * fact that the
-                     * distance of Galatea from a border means that it has no advantage for Mercs
-                     * over border
+                    /* Though the rules do not state these modifiers are mutually exclusive, the fact that the
+                     * distance of Galatea from a border means that it has no advantage for Mercs over border
                      * worlds. Common sense dictates that worlds with hiring halls should not be
                      * subject to the -1 for backwater/interior.
                      */
@@ -653,32 +594,7 @@ public class AtbMonthlyContractMarket implements IContractMarket {
         contract.setEnemyQuality(getQualityRating(Compute.d6(2) + mod));
     }
 
-    protected SkillLevel getSkillRating(int roll) {
-        if (roll <= 5) {
-            return SkillLevel.GREEN;
-        } else if (roll <= 9) {
-            return SkillLevel.REGULAR;
-        } else if (roll <= 11) {
-            return SkillLevel.VETERAN;
-        } else {
-            return SkillLevel.ELITE;
-        }
-    }
-
-    protected int getQualityRating(int roll) {
-        if (roll <= 5) {
-            return IUnitRating.DRAGOON_F;
-        } else if (roll <= 8) {
-            return IUnitRating.DRAGOON_D;
-        } else if (roll <= 10) {
-            return IUnitRating.DRAGOON_C;
-        } else if (roll == 11) {
-            return IUnitRating.DRAGOON_B;
-        } else {
-            return IUnitRating.DRAGOON_A;
-        }
-    }
-
+    @Override
     protected void setAtBContractClauses(AtBContract contract, int unitRatingMod, Campaign campaign) {
         ClauseMods mods = new ClauseMods();
         clauseMods.put(contract.getId(), mods);
@@ -789,160 +705,5 @@ public class AtbMonthlyContractMarket implements IContractMarket {
                 campaign.getCampaignOptions().getContractMaxSalvagePercentage());
         rollSupportClause(contract, mods.mods[CLAUSE_SUPPORT]);
         rollTransportClause(contract, mods.mods[CLAUSE_TRANSPORT]);
-    }
-
-    private void rollCommandClause(final Contract contract, final int modifier) {
-        final int roll = Compute.d6(2) + modifier;
-        if (roll < 3) {
-            contract.setCommandRights(ContractCommandRights.INTEGRATED);
-        } else if (roll < 8) {
-            contract.setCommandRights(ContractCommandRights.HOUSE);
-        } else if (roll < 12) {
-            contract.setCommandRights(ContractCommandRights.LIAISON);
-        } else {
-            contract.setCommandRights(ContractCommandRights.INDEPENDENT);
-        }
-    }
-
-    private void rollSalvageClause(AtBContract contract, int mod, int contractMaxSalvagePercentage) {
-        contract.setSalvageExchange(false);
-        int roll = Math.min(Compute.d6(2) + mod, 13);
-        if (roll < 2) {
-            contract.setSalvagePct(0);
-        } else if (roll < 4) {
-            contract.setSalvageExchange(true);
-            int r;
-            do {
-                r = Compute.d6(2);
-            } while (r < 4);
-            contract.setSalvagePct(Math.min((r - 3) * 10, contractMaxSalvagePercentage));
-        } else {
-            contract.setSalvagePct(Math.min((roll - 3) * 10, contractMaxSalvagePercentage));
-        }
-    }
-
-    private void rollSupportClause(AtBContract contract, int mod) {
-        int roll = Compute.d6(2) + mod;
-        contract.setStraightSupport(0);
-        contract.setBattleLossComp(0);
-        if (roll < 3) {
-            contract.setStraightSupport(0);
-        } else if (roll < 8) {
-            contract.setStraightSupport((roll - 2) * 20);
-        } else if (roll == 8) {
-            contract.setBattleLossComp(10);
-        } else {
-            contract.setBattleLossComp(Math.min((roll - 8) * 20, 100));
-        }
-    }
-
-    private void rollTransportClause(AtBContract contract, int mod) {
-        int roll = Compute.d6(2) + mod;
-        if (roll < 2) {
-            contract.setTransportComp(0);
-        } else if (roll < 6) {
-            contract.setTransportComp((20 + (roll - 2) * 5));
-        } else if (roll < 10) {
-            contract.setTransportComp((45 + (roll - 6) * 5));
-        } else {
-            contract.setTransportComp(100);
-        }
-    }
-
-    @Override
-    public void writeToXML(final PrintWriter pw, int indent) {
-        MHQXMLUtility.writeSimpleXMLOpenTag(pw, indent++, "contractMarket");
-        MHQXMLUtility.writeSimpleXMLTag(pw, indent, "lastId", lastId);
-        for (final Contract contract : contracts) {
-            contract.writeToXML(pw, indent);
-        }
-
-        for (final Integer key : clauseMods.keySet()) {
-            if (!contractIds.containsKey(key)) {
-                continue;
-            }
-
-            MHQXMLUtility.writeSimpleXMLOpenTag(pw, indent++, "clauseMods", "id", key);
-            MHQXMLUtility.writeSimpleXMLTag(pw, indent, "mods", clauseMods.get(key).mods);
-            MHQXMLUtility.writeSimpleXMLTag(pw, indent, "rerollsUsed", clauseMods.get(key).rerollsUsed);
-            MHQXMLUtility.writeSimpleXMLCloseTag(pw, --indent, "clauseMods");
-        }
-        MHQXMLUtility.writeSimpleXMLCloseTag(pw, --indent, "contractMarket");
-    }
-
-    public static AtbMonthlyContractMarket generateInstanceFromXML(Node wn, Campaign c, Version version) {
-        AtbMonthlyContractMarket retVal = null;
-
-        try {
-            // Instantiate the correct child class, and call its parsing function.
-            retVal = new AtbMonthlyContractMarket();
-
-            // Okay, now load Part-specific fields!
-            NodeList nl = wn.getChildNodes();
-
-            // Loop through the nodes and load our contract offers
-            for (int x = 0; x < nl.getLength(); x++) {
-                Node wn2 = nl.item(x);
-
-                // If it's not an element node, we ignore it.
-                if (wn2.getNodeType() != Node.ELEMENT_NODE) {
-                    continue;
-                }
-
-                if (wn2.getNodeName().equalsIgnoreCase("lastId")) {
-                    retVal.lastId = Integer.parseInt(wn2.getTextContent());
-                } else if (wn2.getNodeName().equalsIgnoreCase("mission")) {
-                    Mission m = Mission.generateInstanceFromXML(wn2, c, version);
-
-                    if (m instanceof Contract) {
-                        retVal.contracts.add((Contract) m);
-                        retVal.contractIds.put(m.getId(), (Contract) m);
-                    }
-                } else if (wn2.getNodeName().equalsIgnoreCase("clauseMods")) {
-                    int key = Integer.parseInt(wn2.getAttributes().getNamedItem("id").getTextContent());
-                    ClauseMods cm = new ClauseMods();
-                    NodeList nl2 = wn2.getChildNodes();
-                    for (int i = 0; i < nl2.getLength(); i++) {
-                        Node wn3 = nl2.item(i);
-                        if (wn3.getNodeName().equalsIgnoreCase("mods")) {
-                            String[] s = wn3.getTextContent().split(",");
-                            for (int j = 0; j < s.length; j++) {
-                                cm.mods[j] = Integer.parseInt(s[j]);
-                            }
-                        } else if (wn3.getNodeName().equalsIgnoreCase("rerollsUsed")) {
-                            String[] s = wn3.getTextContent().split(",");
-                            for (int j = 0; j < s.length; j++) {
-                                cm.rerollsUsed[j] = Integer.parseInt(s[j]);
-                            }
-                        }
-                    }
-                    retVal.clauseMods.put(key, cm);
-                }
-            }
-
-            // Restore any parent contract references
-            for (Contract contract : retVal.contracts) {
-                if (contract instanceof AtBContract atbContract) {
-                    atbContract.restore(c);
-                }
-            }
-        } catch (Exception ex) {
-            // Errrr, apparently either the class name was invalid...
-            // Or the listed name doesn't exist.
-            // Doh!
-            logger.error("", ex);
-        }
-
-        return retVal;
-    }
-
-    /*
-     * Keep track of how many rerolls remain for each contract clause
-     * based on the admin's negotiation skill. Also track bonuses, as
-     * the random clause bonuses should be persistent.
-     */
-    public static class ClauseMods {
-        public int[] rerollsUsed = { 0, 0, 0, 0 };
-        public int[] mods = { 0, 0, 0, 0 };
     }
 }
